Updates
-Change ImproperUserIdError to UserIdError
-Move security request error restructuring to security result
-doc and minor updates.
