+++ pyracf/__init__.py
@@ -2,7 +2,7 @@
 from .access.access_admin import AccessAdmin
 from .common.add_operation_error import AddOperationError
 from .common.alter_operation_error import AlterOperationError
-from .common.improper_userid_error import ImproperUserIdError
+from .common.improper_userid_error import UserIdError
 from .common.null_response_error import NullResponseError
 from .common.security_request_error import SecurityRequestError
 from .common.segment_error import SegmentError
