+++ tests/common/test_common_constants.py
@@ -119,7 +119,7 @@ TEST_EXTRACT_RESOURCE_PRECHECK_AS_SQUIDWRD_LOG = get_sample(
 
 TEST_NULL_RESPONSE_PRECHECK_TEXT = (
     "(NullResponseError) Security request made to IRRSMO00 failed."
-    + "\nSAF Return Code: 8 / RACF Return Code: 200 / RACF Reason Code: 16"
+    + "\nSAF Return Code: 8 \nRACF Return Code: 200 \nRACF Reason Code: 16"
     + "\n\nCheck to see if the proper RACF permissions are in place.\n"
     + "For `set` or `alter` functions, you must have at least READ "
     + "access to `IRR.IRRSMO00.PRECHECK` in the `XFACILIT` class."
@@ -127,7 +127,7 @@ TEST_NULL_RESPONSE_PRECHECK_TEXT = (
 
 TEST_NULL_RESPONSE_SURROGAT_TEXT = (
     "(NullResponseError) Security request made to IRRSMO00 failed."
-    + "\nSAF Return Code: 8 / RACF Return Code: 200 / RACF Reason Code: 8"
+    + "\nSAF Return Code: 8 \nRACF Return Code: 200 \nRACF Reason Code: 8"
     + "\n\nCheck to see if the proper RACF permissions are in place.\n"
     + "For the `run_as_userid` feature, you must have at least UPDATE"
     + " access to `ESWIFT.IRRSMO00` in the `SURROGAT` class."
@@ -135,7 +135,7 @@ TEST_NULL_RESPONSE_SURROGAT_TEXT = (
 
 TEST_NULL_RESPONSE_GENERIC_TEXT = (
     "(NullResponseError) Security request made to IRRSMO00 failed."
-    + "\nSAF Return Code: 8 / RACF Return Code: 2000 / RACF Reason Code: 20"
+    + "\nSAF Return Code: 8 \nRACF Return Code: 2000 \nRACF Reason Code: 20"
     + "\n\nPlease check the specified return and reason codes against"
     + " the IRRSMO00 documented return and reason codes for more information"
     + " about this error.\n"
